minor: rename 'tokenizer_internal' -> 'tokenizer_info_impl'
